requirements.txt
# Core requirements
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==20.1.0

# Speech-to-Text
faster-whisper==0.9.0
torch==2.2.0
soundfile==0.12.1
requests==2.31.0

# Text-to-Speech
elevenlabs==0.2.6
gtts==2.3.2

# AI Chat
openai==0.27.8

# NLP Utilities
nltk==3.8.1
langdetect==1.0.9

# Additional dependencies from other files
pydub==0.25.1  # For stt_google.py
google-cloud-speech==2.21.0  # For stt_google.py
google-cloud-translate==2.0.1  # For stt_google.py
vapi-python # For create_assistant.py